IDE workaround
* apparently, ifdef'ed includes are *always* used by the IDE to
"require" library files.  Conditionally-included NeoHWSerial.h seems to
force linking in NeoHWSerial, even when the condition is FALSE.  As a
result, TRUE conditional references to HardwareSerial (via 'Serial')
would conflict with the incorrectly linked NeoHWSerial.  Simply
commenting out the NeoHWSerial include, in a conditional section that
was FALSE, removes the "require" of that library, and HardwareSerial
links in correctly.  Workaround is to always use NeoSerial instead of
Serial.  Sigh.