read returns default fix when none available
* user was causing underrun and getting a fix that was !is_safe().
Returning an empty fix makes it painfully obvious that nothing was
available.
* read() return type is no longer a reference, but as expected, the
caller's instance from the LHS of the assignment is passed in to be
initialized.  That is, this does not require more static RAM for
internal use.
* user feedback suggests emphasizing that the SD card should be flushed
or closed before power down or extraction.  Comment block added to
NMEASDlog.ino