Use pre-defined SERIAL_PORT_HARDWARE_OPEN
*  Default to using AltSoftSerial instead of NeoSWSerial
*  Use the "secret" Arduino symbol that identifies an "open" serial
port, if available.  This is usually `Serial1`, but could occasionally
be `Serial`.  See the platform's variant.h file for definition.  In a
funny coincidence, the comment says that"a GPS module would normally
connect to SERIAL_PORT_HARDWARE_OPEN, the first hardware serial port
whose RX/TX pins are not dedicated to another use."  -_-